requirements.txt
discord.py
psycopg2-binary
pandas
matplotlib
requests

openpyxl
